+++ code/features/aaron.py
@@ -1,0 +1,17 @@
+from database_conn import DB_conn
+
+def aaron_feature1():
+    """
+    Given a pokemon, get a list of pokemon moves (in general) that could be super effective against  it? (pokemon.csv with AllMoves) -- subquery 
+    """
+
+def aaron_feature2():
+    """
+    Given a pokemon and an opponent pokemon, get a list of super effective moves the pokemon can learn against the opponent pokemon
+    """
+    print("Aaron feature 2")
+
+__functions__ = {
+    "aaron_feature1": aaron_feature1,
+    "aaron_feature2": aaron_feature2
+}
