+++ code/namespaces.py
