--- features/aaron.py
@@ -1,10 +1,0 @@
-def aaron_feature1():
-    print("Aaron feature 1")
-
-def aaron_feature2():
-    print("Aaron feature 2")
-
-__functions__ = {
-    "aaron_feature1": aaron_feature1,
-    "aaron_feature2": aaron_feature2
-}
