+++ other_stuff/db-setup.sql
@@ -1,0 +1,7 @@
+DROP DATABASE IF EXISTS pokemondb;
+DROP SCHEMA IF EXISTS pokemon;
+DROP USER IF EXISTS ash;
+
+CREATE USER ash WITH PASSWORD 'ketchum';
+CREATE DATABASE pokemondb;
+GRANT ALL PRIVILEGES ON DATABASE pokemondb TO ash;
