--- other_stuff/pokeDB_setup.sql
@@ -1,3 +1,0 @@
-CREATE USER ash WITH PASSWORD 'ketchum';
-CREATE DATABASE pokemon;
-GRANT ALL PRIVILEGES ON DATABASE pokemon TO ash;
