+++ other_stuff/schema.sql
@@ -1,0 +1,59 @@
+DROP SCHEMA IF EXISTS pokemon;
+CREATE SCHEMA pokemon AUTHORIZATION ash;
+
+CREATE TABLE tbl_allmoves(
+    move_id                 SMALLINT      UNIQUE,
+    name                    VARCHAR(29)   PRIMARY KEY,
+    type                    VARCHAR(8),
+    category                VARCHAR(8),
+    effect                  VARCHAR(255),
+    power                   SMALLINT,
+    acc                     SMALLINT,
+    pp                      SMALLINT,
+    tm                      CHAR(5),
+    prob_second_effect      SMALLINT,
+    gen                     SMALLINT
+);
+
+CREATE TABLE tbl_pokemon(
+    against_bug       REAL,
+    against_dark      REAL,
+    against_dragon    REAL,
+    against_electric  REAL,
+    against_fairy     REAL,
+    against_fight     REAL,
+    against_fire      REAL,
+    against_flying    REAL,
+    against_ghost     REAL,
+    against_grass     REAL,
+    against_ground    REAL,
+    against_ice       REAL,
+    against_normal    REAL,
+    against_poison    REAL,
+    against_psychic   REAL,
+    against_rock      REAL,
+    against_steel     REAL,
+    against_water     REAL,
+    attack            SMALLINT,
+    base_egg_steps    SMALLINT,
+    base_happiness    SMALLINT,
+    base_total        SMALLINT,
+    capture_rate      SMALLINT,
+    classification    VARCHAR(63),
+    defense           SMALLINT,
+    experience_growth INTEGER,
+    height_m          REAL,
+    hp                SMALLINT,
+    japanese_name     VARCHAR(63)  COLLATE "ja-JP-x-icu",
+    name              VARCHAR(15)  UNIQUE,
+    percentage_male   REAL,
+    pokedex_number    SMALLINT     PRIMARY KEY,
+    sp_attack         SMALLINT,
+    sp_defense        SMALLINT,
+    speed             SMALLINT,
+    type1             VARCHAR(15),
+    type2             VARCHAR(15),
+    weight_kg         REAL,
+    generation        SMALLINT,
+    is_legendary      BOOLEAN
+);
